disco: keep transaction metadata within 40 bytes
Narrow payload_sz to ushort and group the source fields beside it so
fd_txn_p_t occupies 4992 bytes instead of 5056.  fd_txn_e_t also loses
one cache line, and the default replay transaction pool saves 4 MiB.
Keep first_seen_nanos in fd_txn_p_t for the existing telemetry path.

Assert the payload bound and compact footprint, explicitly narrow the
bounded producers, and update the scheduler footprint expectations.

Adapted from FireBAM commit 7d21e18e6ad231dbdc1d6a8c8bd48b14793a925d.

+++ src/choreo/tower/fd_tower.c
@@ -1371,7 +1371,7 @@ fd_tower_to_vote_txn( fd_tower_t const *    tower,
     ix_accs[1] = 1; /* vote authority */
     program_id = 3; /* vote program */
   }
-  vote_txn->payload_sz = fd_txn_add_instr( txn_meta_out, txn_out, program_id, ix_accs, 2, vote_ix_buf, vote_ix_sz );
+  vote_txn->payload_sz = (ushort)fd_txn_add_instr( txn_meta_out, txn_out, program_id, ix_accs, 2, vote_ix_buf, vote_ix_sz );
 }
 
 int
