+++ src/disco/pack/test_pack.c
@@ -203,7 +203,7 @@ make_transaction1( fd_txn_p_t * txnp,
     }
   }
 
-  txnp->payload_sz = (ulong)(p-p_base);
+  txnp->payload_sz = (ushort)(p-p_base);
   uint flags;
   fd_ulong_store_if( !!priority_fees, priority_fees, (rewards_per_cu * compute + 999999UL)/1000000UL );
   fd_ulong_store_if( !!pack_cost_estimate, pack_cost_estimate, fd_pack_compute_cost( TXN( txnp ), txnp->payload, &flags, NULL, NULL, NULL, NULL, NULL ) );
@@ -226,7 +226,7 @@ make_vote_transaction1( fd_txn_p_t * txnp,
                         ulong        i ) {
   uchar * p = txnp->payload;
   fd_memcpy( p, sample_vote, sample_vote_sz );
-  txnp->payload_sz = sample_vote_sz;
+  txnp->payload_sz = (ushort)sample_vote_sz;
 
   /* Make signature and the two writable accounts unique */
   p[ 0x01+(i%8) ] = (uchar)(p[ 0x01+(i%8) ] + 1UL + (i/8));
@@ -327,7 +327,7 @@ make_nonce_transaction1( fd_txn_p_t * txnp,
   *(ptrs[1]) = 3; fd_memcpy( ptrs[1]+1, &rewards_per_cu, sizeof(ulong) );
   *(ptrs[2]) = 4; fd_memcpy( ptrs[2]+1, &loaded_data_sz, sizeof(uint)  );
 
-  txnp->payload_sz = (ulong)(p-p_base);
+  txnp->payload_sz = (ushort)(p-p_base);
 }
 
 static void
@@ -751,7 +751,7 @@ performance_test2( void ) {
       for( ulong i=0UL; i<1024UL; i++ ) {
         fd_txn_e_t * slot      = fd_pack_insert_txn_init( pack );
         fd_txn_t *   txn       = (fd_txn_t *)txn_scratch[ i ];
-        slot->txnp->payload_sz = payload_sz[ i ];
+        slot->txnp->payload_sz = (ushort)payload_sz[ i ];
         fd_memcpy( slot->txnp->payload, payload_scratch[ i ], payload_sz[ i ]                                                );
         fd_memcpy( TXN(slot->txnp),     txn,                  fd_txn_footprint( txn->instr_cnt, txn->addr_table_lookup_cnt ) );
         ulong _deleted;
@@ -837,7 +837,7 @@ void performance_test( int extra_bench ) {
         memcpy( payload_scratch[j&1]+1UL, &j, sizeof(ulong) );
         fd_txn_e_t * slot       = fd_pack_insert_txn_init( pack );
         fd_txn_t *   txn        = (fd_txn_t*) txn_scratch[ j&1 ];
-        slot->txnp->payload_sz  = payload_sz[ j&1 ];
+        slot->txnp->payload_sz  = (ushort)payload_sz[ j&1 ];
         fd_memcpy( slot->txnp->payload, payload_scratch[ j&1 ], payload_sz[ j&1 ]                                              );
         fd_memcpy( TXN(slot->txnp),     txn,                    fd_txn_footprint( txn->instr_cnt, txn->addr_table_lookup_cnt ) );
         fd_pack_insert_txn_cancel( pack, slot );
@@ -849,7 +849,7 @@ void performance_test( int extra_bench ) {
         memcpy( payload_scratch[j&1]+1UL, &j, sizeof(ulong) );
         fd_txn_e_t * slot       = fd_pack_insert_txn_init( pack );
         fd_txn_t *   txn        = (fd_txn_t*) txn_scratch[ j&1 ];
-        slot->txnp->payload_sz  = payload_sz[ j&1 ];
+        slot->txnp->payload_sz  = (ushort)payload_sz[ j&1 ];
         fd_memcpy( slot->txnp->payload, payload_scratch[ j&1 ], payload_sz[ j&1 ]                                              );
         fd_memcpy( TXN(slot->txnp),     txn,                    fd_txn_footprint( txn->instr_cnt, txn->addr_table_lookup_cnt ) );
 
@@ -895,7 +895,7 @@ void performance_test( int extra_bench ) {
         memcpy( payload_scratch[j&1]+1UL, &j, sizeof(ulong) );
         fd_txn_e_t * slot       = fd_pack_insert_txn_init( pack );
         fd_txn_t *   txn        = (fd_txn_t*) txn_scratch[ j&1 ];
-        slot->txnp->payload_sz  = payload_sz[ j&1 ];
+        slot->txnp->payload_sz  = (ushort)payload_sz[ j&1 ];
         fd_memcpy( slot->txnp->payload, payload_scratch[ j&1 ], payload_sz[ j&1 ]                                              );
         fd_memcpy( TXN(slot->txnp),     txn,                    fd_txn_footprint( txn->instr_cnt, txn->addr_table_lookup_cnt ) );
 
@@ -924,7 +924,7 @@ void performance_test( int extra_bench ) {
         memcpy( payload_scratch[j&1]+1UL, &j, sizeof(ulong) );
         fd_txn_e_t * slot       = fd_pack_insert_txn_init( pack );
         fd_txn_t *   txn        = (fd_txn_t*) txn_scratch[ j&1 ];
-        slot->txnp->payload_sz  = payload_sz[ j&1 ];
+        slot->txnp->payload_sz  = (ushort)payload_sz[ j&1 ];
         fd_memcpy( slot->txnp->payload, payload_scratch[ j&1 ], payload_sz[ j&1 ]                                              );
         fd_memcpy( TXN(slot->txnp),     txn,                    fd_txn_footprint( txn->instr_cnt, txn->addr_table_lookup_cnt ) );
 
@@ -1002,7 +1002,7 @@ void performance_end_block( void ) {
         }
         fd_txn_e_t * slot      = fd_pack_insert_txn_init( pack );
         fd_txn_t *   txn       = (fd_txn_t*) txn_scratch[ 0UL ];
-        slot->txnp->payload_sz = payload_sz[ 0UL ];
+        slot->txnp->payload_sz = (ushort)payload_sz[ 0UL ];
         fd_memcpy( slot->txnp->payload, payload_scratch[ 0UL ], payload_sz[ 0UL ]                                              );
         fd_memcpy( TXN(slot->txnp),     txn,                    fd_txn_footprint( txn->instr_cnt, txn->addr_table_lookup_cnt ) );
